adding openai

+++ src/model/infer_response.py
@@ -3,6 +3,6 @@ from pydantic import BaseModel
 class InferResponse(BaseModel):
     provider: str
     model: str
-    latency_ms: int
+    latency_ms: float
     cost_usd: float
     response: str
